readyup|refactor: Split argument parsing out of the run command
Moves run-subcommand argument parsing out of `runCommand.ts` into two modules under `src/run/`. `parseRunArgs.ts` takes the option table, the severity and detail coercion, and `ParsedRunArgs`; `validateRunFlags.ts` takes the output, exclusivity, mode-flag, and checklist-selection constraints. `runCommand.ts` drops from 1076 to 810 lines, and `bin/route.ts` imports `parseRunArgs` from its own module.

`validateFlagConstraints` is renamed `validateRunFlags` so the file takes the name of its main export; every other declaration moves verbatim.

Each module's tests sit beside it in `src/run/__tests__/`, call the module directly, and mock nothing. The constraint matrix is now exercised against `validateRunFlags` with a constraints object instead of a flag array, which adds the `--internal` with `--packages` pairing the old assertions missed. Coverage totals are unchanged.
